requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy>=2.0.40,<2.1
psycopg[binary]==3.2.13
pydantic>=2.12,<3
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt<5
python-multipart==0.0.20
email-validator==2.2.0
pytest==8.3.4
httpx==0.28.1
